model/train_model_improved.py
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn.preprocessing import StandardScaler
from sklearn.metrics import accuracy_score, classification_report
import pickle
import os

def train_and_save_model(output_path: str = 'model.pkl', scaler_path: str = 'scaler.pkl', data_path: str = 'training_data.csv'):
    """
    Trains a Random Forest model.
    Updated: Reduced noise to improve accuracy for demo purposes (~90%+).
    """
    print("Generating synthetic data...")
    data_size = 2000  # Increased data size for better learning
    np.random.seed(42)

    # 1. Generate Features
    data = {
        'attendance': np.random.randint(40, 100, data_size),
        'study_hours': np.random.randint(0, 15, data_size),
        'internal_marks': np.random.randint(10, 100, data_size),
        'assignments_submitted': np.random.randint(0, 10, data_size),
        'activities_participation': np.random.randint(0, 2, data_size)  # 0 = No, 1 = Yes
    }

    df = pd.DataFrame(data)
    
    # 2. Determine Target
    def probabilistic_pass_fail(row):
        # Base score calculation
        score = (row['attendance'] * 0.3) + (row['internal_marks'] * 0.5) + (row['study_hours'] * 3)
        
        if row['assignments_submitted'] < 5:
            score -= 10
        
        if row['activities_participation'] == 1:
            score += 5

        # REDUCED NOISE: Changed std_dev from 8 to 3. 
        # This makes the data easier for the model to predict, boosting accuracy.
        noise = np.random.normal(0, 3) 
        final_score = score + noise

        return 1 if final_score > 65 else 0

    df['performance'] = df.apply(probabilistic_pass_fail, axis=1)

    # Save synthetic data
    data_dir = os.path.dirname(os.path.abspath(data_path))
    if data_dir and not os.path.exists(data_dir):
        os.makedirs(data_dir, exist_ok=True)
    df.to_csv(data_path, index=False)
    print(f"Synthetic data saved to {data_path}")

    # 3. Preprocessing (Scaling)
    X = df.drop('performance', axis=1)
    y = df['performance']

    scaler = StandardScaler()
    X_scaled = scaler.fit_transform(X)

    X_train, X_test, y_train, y_test = train_test_split(X_scaled, y, test_size=0.2, random_state=42)

    # 4. Train Model
    print("Training Random Forest...")
    # Increased estimators to 200 for better stability
    model = RandomForestClassifier(n_estimators=200, max_depth=12, random_state=42)
    model.fit(X_train, y_train)

    # 5. Evaluate
    predictions = model.predict(X_test)
    acc = accuracy_score(y_test, predictions)
    print(f"Model Accuracy: {acc*100:.2f}%")
    print("Classification Report:")
    print(classification_report(y_test, predictions))

    # 6. Save Model AND Scaler
    with open(output_path, 'wb') as f:
        pickle.dump(model, f)
    
    with open(scaler_path, 'wb') as f:
        pickle.dump(scaler, f)

    print(f"Model saved as '{output_path}'")
    print(f"Scaler saved as '{scaler_path}'")

    return model, scaler

if __name__ == "__main__":
    train_and_save_model()