requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
gunicorn
google-generativeai